Minor fixes

+++ include/density/density.h
@@ -100,7 +100,7 @@ private:
   std::vector<float> velocs[dim];                  // size: local_particles
   std::vector<float> density_field;                // size: local_rho_count
   std::vector<long> histogram;                     // size: nb_bins
-  std::vector<std::vector<long>> buckets;          // size: nb_bins
+  std::vector<std::vector<int>> buckets;          // size: nb_bins
   std::vector<int> bits;                           // size: nb_bins
   std::vector<long> index;                         // size: local_particles
   std::vector<float> decompressed[2 * dim];        // size: local_particles
